refactor: Relocate `GsapImpl` to `$lib` for better structure and modular imports
