+++ svelte.config.js
@@ -3,6 +3,9 @@ import adapter from '@sveltejs/adapter-static';
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	kit: {
+		alias: {
+			$lib: 'src/lib',
+		},
 		adapter: adapter({
 			pages: 'build',
 			assets: 'build',
